First split of falco-related code to falco_engine.
Create a standalone class falco_engine that will be the primary
interface for using falco. Its main interfaces are:

 - init(): given a path to a rules file, create a lua state, locate the
   lua code, and load the set of rules in the file.
 - add_output(): add an output channel for alert notifications.
 - handle_event(): given an event that matches one or more rules, format
   an alert message and send it to any output channels.
 - print_stats(): print stats on the rules that matched.

All of the code to implement these was taken from the main falco program
itself. Eventually, falco_engine will be a library and the main falco
program will link with it.