Export output_config from falco_engine.
With the current engine, reading/parsing the main config file falco.yaml
is outside the engine, but the engine needed the type output_config,
defined in configuration.h, so it could be one of the arguments to
falco_engine::add_output.

Move this type to falco_engine and use it in configuration.h, so the
engine can be self-contained.